+++ app/src/main/java/com/kwabenaberko/finito/viewmodel/dto/NoteListItem.kt
@@ -1,7 +1,7 @@
 package com.kwabenaberko.finito.viewmodel.dto
 
 class NoteListItem(
-        val noteId: Int,
+        val noteId: Long,
         val text: String,
         var color: String
 )
